Initial e-commerce demo: Spring Boot backend, static frontend, entities, services, and DB migrations

+++ src/main/java/com/huawei/cloudnative/ecommerce/model/Product.java
@@ -1,4 +1,6 @@
 package com.huawei.cloudnative.ecommerce.model;
 
-public record Product(Long id, String name, String description, double price) {
+import java.math.BigDecimal;
+
+public record Product(Long id, String name, String description, BigDecimal price) {
 }
